// frontend/src/api/generated/model/taskResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { TaskResponseDomainTags } from "./taskResponseDomainTags";
import type { TaskResponseRequiredSkills } from "./taskResponseRequiredSkills";

export interface TaskResponse {
  id: string;
  workspace_id: string;
  organisation_id: string;
  status: string;
  title: string;
  description: string | null;
  task_type: string | null;
  priority: string | null;
  deadline_at: string | null;
  external_ref: string | null;
  idempotency_key: string | null;
  required_skills: TaskResponseRequiredSkills;
  difficulty: number | null;
  domain_tags: TaskResponseDomainTags;
  created_at: string | null;
}